+++ python/rlgym_learn_algos/agent_controller/multi_agent/__init__.py
@@ -1,19 +1,15 @@
 __all__ = [
     "DerivedMultiAgentSubcontrollerConfig",
-    "EnvActionResponse",
-    "EnvActionResponseType",
     "MultiAgentController",
     "MultiAgentControllerConfigModel",
     "MultiAgentSubcontroller",
 ]
 
-from .._rlgym_learn_algos.agent_controller import (
-    EnvActionResponse,
-    EnvActionResponseType,
-)
-from .multi_agent import (
-    DerivedMultiAgentSubcontrollerConfig,
+from .multi_agent_controller import (
     MultiAgentController,
     MultiAgentControllerConfigModel,
+)
+from .multi_agent_subcontroller import (
+    DerivedMultiAgentSubcontrollerConfig,
     MultiAgentSubcontroller,
 )