+++ python/rlgym_learn_algos/util/torch_functions.py
@@ -8,7 +8,7 @@ Description:
 """
 
 import torch
-import torch.nn as nn
+from torch import nn
 from typing_extensions import override